requirements.txt
Flask==3.1.2
gunicorn==25.1.0
numpy==2.2.6
pandas==2.3.3
scikit-learn==1.7.2
scipy==1.15.3
nltk==3.9.2
joblib==1.5.3
pyspellchecker==0.8.4
textstat==0.7.12
regex==2026.1.15

